refactor hook tests to use the github event server
